Cleaned up project.

+++ Project/src/com/radirius/mercury/math/MathUtil.java
@@ -11,7 +11,7 @@ public class MathUtil {
 	/**
 	 * Base method for random number methods.
 	 *
-	 * @return A random double value between minimum and maximum
+	 * Returns A random double value between minimum and maximum
 	 */
 	public static double random(double minimum, double maximum) {
 		return minimum + (Math.random() * (maximum - minimum + 1));
@@ -20,22 +20,22 @@ public class MathUtil {
 	/**
 	 * Base method for random number methods.
 	 *
-	 * @return A random double value between minimum and maximum
+	 * Returns A random double value between minimum and maximum
 	 */
 	public static float random(float minimum, float maximum) {
 		return minimum + (float) (Math.random() * (maximum - minimum + 1));
 	}
 
 	/**
-	 * @return A random boolean. 50-50 chance of true or false.
+	 * Returns A random boolean. 50-50 chance of true or false.
 	 */
 	public static boolean nextBoolean() {
 		return (int) random(0, 20) % 2 == 0;
 	}
 
 	/**
-	 * @param percent Percent chance of true
-	 * @return A boolean that has a percent% chance of being true.
+	 * @param percent Percent chance of true Returns A boolean that has a
+	 *        percent% chance of being true.
 	 */
 	public static boolean chance(int percent) {
 		if (percent > 100)
@@ -48,8 +48,8 @@ public class MathUtil {
 	}
 
 	/**
-	 * @param percent Percent chance of true
-	 * @return A boolean that has a percent% chance of being true.
+	 * @param percent Percent chance of true Returns A boolean that has a
+	 *        percent% chance of being true.
 	 */
 	public static boolean chance(float percent) {
 		if (percent > 1)
@@ -62,14 +62,14 @@ public class MathUtil {
 	}
 
 	/**
-	 * @return An integer value between Integer.MIN_VALUE and Integer.MAX_VALUE.
+	 * Returns An integer value between Integer.MIN_VALUE and Integer.MAX_VALUE.
 	 */
 	public static int nextInt() {
 		return (int) random(Integer.MIN_VALUE, Integer.MAX_VALUE);
 	}
 
 	/**
-	 * @return Either 1 or -1.
+	 * Returns Either 1 or -1.
 	 */
 	public static int negpos() {
 		return nextBoolean() ? 1 : -1;
@@ -83,56 +83,56 @@ public class MathUtil {
 	}
 
 	/**
-	 * @return A floating point value between 0.0 and 1.0
+	 * Returns A floating point value between 0.0 and 1.0
 	 */
 	public static float nextFloat() {
-		return (float) random(0, 100) / 100;
+		return random(0, 100) / 100;
 	}
 
 	/**
-	 * @return A double value between Double.MIN_VALUE and Double.MAX_VALUE.
+	 * Returns A double value between Double.MIN_VALUE and Double.MAX_VALUE.
 	 */
 	public static double nextDouble() {
 		return random(Double.MIN_VALUE, Double.MAX_VALUE);
 	}
 
 	/**
-	 * @return The sine value of an angle.
+	 * Returns The sine value of an angle.
 	 */
 	public static float sin(float angle) {
 		return (float) Math.sin(Math.toRadians(angle));
 	}
 
 	/**
-	 * @return The cosine value of an angle.
+	 * Returns The cosine value of an angle.
 	 */
 	public static float cos(float angle) {
 		return (float) Math.cos(Math.toRadians(angle));
 	}
 
 	/**
-	 * @return The arch-tangent (tan^-1) of x, and y.
+	 * Returns The arch-tangent (tan^-1) of x, and y.
 	 */
 	public static float atan2(float x, float y) {
 		return (float) Math.toDegrees(Math.atan2(y, x));
 	}
 
 	/**
-	 * @return The equivalent of an angle in radians.
+	 * Returns The equivalent of an angle in radians.
 	 */
 	public static double toRadians(double angle) {
 		return angle * (Math.PI / 180);
 	}
 
 	/**
-	 * @return The equivalent of an angle in degrees.
+	 * Returns The equivalent of an angle in degrees.
 	 */
 	public static double toDegrees(double angle) {
 		return angle * (180 / Math.PI);
 	}
 
 	/**
-	 * @return If number is negative, -1, otherwise, 1.
+	 * Returns If number is negative, -1, otherwise, 1.
 	 */
 	public static float negpos(float mult) {
 		if (mult == 0)
